rename testUtil to projectUtil and remove deprecated web3util
